requirements.txt
Jinja2~=3.0
PyYAML~=6.0
snowflake-connector-python>=2.8,<4.0
structlog~=24.1.0

# Supported databases
psycopg==3.2.6
psycopg-binary==3.2.6
pymssql==2.3.4
mysql-connector-python==9.3.0
oracledb==3.1.1